Rename `ensv2` plugin to `unigraph`
