+++ apps/ensindexer/ponder/ponder.config.ts
@@ -16,9 +16,9 @@ logger.info({
 });
 
 // Log warning about dual activation of subgraph and ensv2 plugins
-if (config.plugins.includes(PluginName.Subgraph) && config.plugins.includes(PluginName.ENSv2)) {
+if (config.plugins.includes(PluginName.Subgraph) && config.plugins.includes(PluginName.Unigraph)) {
   logger.warn({
-    msg: `Both the '${PluginName.Subgraph}' and '${PluginName.ENSv2}' plugins are enabled. This results in the availability of both the legacy Subgraph-Compatible GraphQL API (/subgraph) _and_ ENSNode's Omnigraph API (/api/omnigraph), and comes with an associated increase in indexing time.`,
+    msg: `Both the '${PluginName.Subgraph}' and '${PluginName.Unigraph}' plugins are enabled. This results in the availability of both the legacy Subgraph-Compatible GraphQL API (/subgraph) _and_ ENSNode's Omnigraph API (/api/omnigraph), and comes with an associated increase in indexing time.`,
     advice: `If your intent is to have both APIs available in parallel, excellent, otherwise you may benefit from only enabling the plugin for the API you plan to use.`,
   });
 }
