+++ apps/ensindexer/src/plugins/unigraph/handlers/ensv1/BaseRegistrar.ts
@@ -34,7 +34,7 @@ import { getManagedName } from "@/lib/managed-names";
 import { namespaceContract } from "@/lib/plugin-helpers";
 import type { EventWithArgs } from "@/lib/ponder-helpers";
 
-const pluginName = PluginName.ENSv2;
+const pluginName = PluginName.Unigraph;
 
 /**
  * In ENSv1, all BaseRegistrar-derived Registrar contracts (& their controllers) have the ability to
