+++ packages/ensnode-sdk/src/ensindexer/config/types.ts
@@ -16,7 +16,7 @@ export enum PluginName {
   ProtocolAcceleration = "protocol-acceleration",
   Registrars = "registrars",
   TokenScope = "tokenscope",
-  ENSv2 = "ensv2",
+  Unigraph = "unigraph",
 }
 
 /**
