+++ packages/ensnode-sdk/src/omnigraph-api/prerequisites.ts
@@ -7,12 +7,12 @@ import type { PrerequisiteResult } from "../shared/prerequisites";
  * Check if provided EnsIndexerPublicConfig supports the Omnigraph API.
  */
 export function hasOmnigraphApiConfigSupport(config: EnsIndexerPublicConfig): PrerequisiteResult {
-  const supported = config.plugins.includes(PluginName.ENSv2);
+  const supported = config.plugins.includes(PluginName.Unigraph);
   if (supported) return { supported };
 
   return {
     supported: false,
-    reason: `The connected ENSNode's Config must have the '${PluginName.ENSv2}' plugin enabled.`,
+    reason: `The connected ENSNode's Config must have the '${PluginName.Unigraph}' plugin enabled.`,
   };
 }
